==> 3.2 Estadísticas totales de quintetos.ipynb <==
import pandas as pd
import os
from funciones_auxiliares import *
import warnings
warnings.filterwarnings('ignore')

# 3.2 *Estadísticas totales de quintetos*

Las *estadísticas totales de quinteto*, también conocidas como estadísticas acumuladas de quinteto, es la información que representa las estadísticas obtenidas durante por un conjunto de cinco jugadores en un periodo de tiempo como puede ser un número de partidos específicos, una temporada o varias temporadas cuando estos están juntos en pista. En este notebook se buscará dar explicación de manera entendible y visualizada a todo el procesamiento realizado para la construcción de estas estadísticas a través de los siguientes puntos:
* **1. Base teórica de la construcción de las estadísticas.** Se realizará la explicación teórica de los pasos realizados para la construcción de este tipo de estadísticas.
* **2. Procesamiento de los datos.** Se realizará la explicación guiada del procesamiento de los datos para la construcción de las estadísticas totales.

A través de esos puntos se realizará la comprensión y entendimiento del trabajo realizado para la construcción de las *estadísticas totales*.

## 1. Base teórica de la construcción de las estadísticas

Como ha sido comentado en la introducción de este notebook, las estadísticas totales de quintos, es la forma de representación creada para la representación de las estadísticas que se producen en pista cuando cinco jugadores de un mismo equipo se encuentran en pista. Para la construcción de este tipo de estadística, es necesario poder haber descargado toda la información referente a los *play by play* de los equipos durante un periodo de tiempo. Como ha sido explicado a través del apartado [Descarga del *Play by play*.](2.4%20Play%20by%20play), han sido descargados todos los datos referentes al *play by play* de los encuentros disputados entre las temporadas *2019-2020*, *2020-2021', *2021-2022* y *2022-2023* los cuales serán utilizados para el cálculo de las *estadisticas totales de quintetos* para cada uno de los equipos en cada una de las temporadas.

Teniendo esto presente, la construcción de las *estadísticas totales de los quintetos* de un equipo en una temporada, está compuesta por dos procesamientos de los datos. Por un lado, es necesario el procesamiento del *play by play* en donde se realiza la extracción de los cinco jugadores de un mismo equipo y la construcción de diferentes estadísticas acumuladas que son realizadas por los jugadores que conforman el quinteto o los jugadores contrarios que se enfrenta el quinteto. La estadísticas que son construidas para cada quinteto son las siguientes:
* *T2A.* Número de tiro de dos puntos anotados por los jugadores que conforman el quinteto.
* *T2L.* Número de tiro de dos puntos lanzados por los jugadores que conforman el quinteto.
* *T3A.* Número de tiro de tres puntos anotados por los jugadores que conforman el quinteto.
* *T3L.* Número de tiro de tres puntos lanzados por los jugadores que conforman el quinteto.
* *TLA.* Número de tiro libres anotados por los jugadores que conforman el quinteto.
* *TLL.* Número de tiro libres lanzados por los jugadores que conforman el quinteto.
* *RO.* Número de rebotes ofensivos capturados por los jugadores que conforman el quinteto.
* *RD.* Número de rebotes defensivos capturados por los jugadores que conforman el quinteto.
* *STL.* Número de robos realizados por los jugadores que conforman el quinteto.
* *TOV.* Número de pérdidas realizadas por los jugadores que conforman el quinteto.
* *AST.* Número de asistencias realizadas por los jugadores que conforman el quinteto.
* *BLK.* Número de tapones realizados por los jugadores que conforman el quinteto.
* *T2A_R.* Número de tiro de dos puntos anotados por los jugadores a los que se enfrenta el quinteto.
* *T2L_R.* Número de tiro de dos puntos lanzados por los jugadores a los que se enfrenta el quinteto.
* *T3A_R.* Número de tiro de tres puntos anotados por los jugadores a los que se enfrenta el quinteto.
* *T3L_R.* Número de tiro de tres puntos lanzados por los jugadores a los que se enfrenta el quinteto.
* *TLA_R.* Número de tiro libres anotados por los jugadores a los que se enfrenta el quinteto.
* *TLL_R.* Número de tiro libres lanzados por los jugadores a los que se enfrenta el quinteto.
* *RO_R.* Número de rebotes ofensivos capturados por los jugadores a los que se enfrentaa los que se enfrenta el quinteto.
* *RD_R.* Número de rebotes defensivos capturados por los jugadores que conforman el quinteto.
* *STL_R.* Número de robos realizados por los jugadores a los que se enfrenta el quinteto.
* *TOV_R.* Número de pérdidas realizadas por los jugadores a los que se enfrenta el quinteto
* *AST_R.* Número de asistencias realizadas por los jugadores a los que se enfrenta el quinteto.
* *BLK_R.* Número de tapones realizados por los jugadores a los que se enfrenta el quinteto.
* *TIEMPO_PISTA.* Número de segundos disputados por el quinteto.

Tras la construcción de estas estadísticas, y como segundo procesamiento, se realizará el sumatorio de todos los apartados estadísticos del quinteto para cada equipo y en cada temporada teniendo como resultado todos los apartados estadísticos de los quintetos de un equipo a lo largo de una temporada que hayan participado en la competición.

Para la construcción de las *estadísticas totales de quinteto* ha sido desarrollado el script de código [03_Procesamiento_PBP](03_Procesamiento_PBP.py) el cual permite la lectura de todos los *play by play* de los encuentros disputados por los equipos y se procede a la construcción de este tipo de estadística.

# 2. Procesamiento de los datos

Como ha sido explicado en el apartado anterior, para la construcción de las estadísticas totales, es necesario realizar primero el procesamiento del *play by play* con el fin de construir las estadísticas de cada uno de los quintetos en los partidos que haya disputado. Para ello, se realiza la carga del primer partido disputado en la jornada 1 de la temporada *2022-2023*:

# Ruta para acceder al fichero del play by play del partido y lectura del partido
partido_2022 = 'E:/TFM/02. Datos/ACB/JORNADAS/2022/Jornada 1/20220928_Bàsquet Girona_VS_Real Madrid/PBP_Bàsquet Girona_VS_Real Madrid.csv'
df_partido_2022 = pd.read_csv(partido_2022, sep=';')

print("################################################################################")
print("# PLAY BY PLAY DESCARGADO")
print("################################################################################")
print(df_partido_2022.head(20).to_string(index=False, col_space=8))

Como ha sido comentado, el procesamiento del *play by play* consiste en la extracción de cada combinación de cinco jugadores que compartieron pista para un mismo equipo durante el encuentro y la extracción de las diferentes estadísticas anteriormente comentadas. Se muestra a continuación el procesamiento del *play by play* de ambos equipos:

print("################################################################################")
print("# PLAY BY PLAY PROCESADO Real Madrid")
print("################################################################################")

pbp_procesado = procesar_pbp(df_partido_2022,'Real Madrid')
print(pbp_procesado.to_string(index=False, col_space=1))

print("\n################################################################################")
print("# PLAY BY PLAY PROCESADO Bàsquet Girona")
print("################################################################################")

pbp_procesado = procesar_pbp(df_partido_2022,'Bàsquet Girona')
print(pbp_procesado.to_string(index=False, col_space=1))

Tras el procesamiento de cada *play by play* para cada uno de los equipos en cada una de las temporadas que hayan participado en la liga ACB han sido construidas las estadísticas totales de los quintetos de estos equipos. Se muestra a continuación las *estadísticas totales de quintetos* del Real Madrid durante la temporada *2022-2023* tras la realización del sumatorio de estadísticas:

# Ruta para acceder al fichero de estadísticas totales del Real Madrid y lectura del fichero
fichero_total_2022 = 'E:/TFM/02. Datos/ACB/ESTADISTICAS/2022/Real Madrid/03_ESTADISTICAS_QUINTETO_EQUIPO.csv'
df_estadisticas_totales_quinteto = pd.read_csv(fichero_total_2022, sep=';')

print("################################################################################")
print("# ESTADISTICAS TOTALES DE QUINTETOS DE REAL MADRID EN LA TEMPORADA 2022-2023")
print("################################################################################")

pbp_procesado = procesar_pbp(df_partido_2022,'Real Madrid')
print(df_estadisticas_totales_quinteto.sort_values(by='TIEMPO_PISTA', ascending=False).head(20).to_string(index=False, col_space=1))
